Refactor transaction tests to different files
