+++ __tests__/pages/wallet/util/walletElementSelectors.tsx
@@ -1,0 +1,5 @@
+import { screen } from "@testing-library/react"
+
+export function getTransactionHistory(): HTMLElement {
+  return screen.getByLabelText("wallet-transaction-history")
+}
